added the logic for reading the data from the soil sensor. In theory works.
MX_GPIO_Init (the pa7 is initialized output_pp not mode_analog), HAL_ADC_MspInit e HAL_ADC_MspDeInit (added the gpio for the adc_in4) in adc.c, Read_Channel in adc_if.c removed static. probably to add the #define for the soil sensor in this modification ???

in this moment the percentage is not between 0-100% (in fact when in water the soil perc is around 200).

changed the Payload_Formatter_TTN.txt and the My_Adv_Data_Init